+++ include/kernel.h
@@ -22,32 +22,37 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #define THREAD_HZ 100
 
-#define __DEBUG__ 1
+#define __DEBUG__ 0
 
 #ifdef __DEBUG__
-    #define MAX_LOG_LEN 16
+    #define MAX_LOG_LEN 30
+
+    // Just a number you can set for debugging purposes
+    extern int logged_num;
 
     extern const char* call_log[MAX_LOG_LEN];
     extern int log_index;
     extern int last_call_finished;
 
     static inline void start_call_log(const char* func_name) {
-        int idx = log_index % MAX_LOG_LEN;
-        log_index++;
-
-        call_log[idx] = func_name;
+        log_index = (log_index + 1) % MAX_LOG_LEN;
+        call_log[log_index] = func_name;
         last_call_finished = 0;
     }
 
     static inline void end_call_log(void* res) {
+        (void) res;
         last_call_finished = 1;
     }
 
     #define LOG_CALL() \
         start_call_log(__func__); \
         int _sentinel __attribute__((cleanup(end_call_log))) = 0
+
+    #define LOG_NUM(n) (logged_num = (n))
 #else
     #define LOG_CALL()
+    #define LOG_NUM()
 #endif
 
 // Reorders the machine code to place these accordingly